Rename ValidatorDbApiImplSlick to SlickApi
